added textures

+++ src/EBO.cpp
@@ -1,6 +1,6 @@
 #include "headers/EBO.h"
 
-EBO::EBO(unsigned int* indices, size_t size) {
+EBO::EBO(unsigned int* indices, unsigned int size) {
     glGenBuffers(1, &ID);
     glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, ID);
     glBufferData(GL_ELEMENT_ARRAY_BUFFER, size, indices, GL_STATIC_DRAW);
